Reorganize C and C++ implementations
