+++ cplusplus/cmake/install.cmake
@@ -8,8 +8,7 @@ if(AUTOGRAPH_INSTALL)
     INCLUDES
     DESTINATION include)
 
-  install(FILES ${CMAKE_CURRENT_SOURCE_DIR}/include/autograph.h
-          DESTINATION include)
+  install(DIRECTORY ${CMAKE_CURRENT_SOURCE_DIR}/include/ DESTINATION include)
 
   install(
     EXPORT autograph-targets
