+++ cplusplus/cmake/sources.cmake
@@ -8,4 +8,5 @@ string(REPLACE "\n" ";" SODIUM_SOURCES ${SODIUM_SOURCES})
 file(GLOB_RECURSE AUTOGRAPH_C_SOURCES ${CMAKE_CURRENT_SOURCE_DIR}/src/*.c)
 file(GLOB_RECURSE AUTOGRAPH_CPP_SOURCES ${CMAKE_CURRENT_SOURCE_DIR}/src/*.cpp)
 
-set(AUTOGRAPH_SOURCES ${SODIUM_SOURCES} ${AUTOGRAPH_C_SOURCES} ${AUTOGRAPH_CPP_SOURCES})
+set(AUTOGRAPH_SOURCES ${SODIUM_SOURCES} ${AUTOGRAPH_C_SOURCES}
+                      ${AUTOGRAPH_CPP_SOURCES})
