+++ cplusplus/include/autograph/init.h
@@ -1,0 +1,20 @@
+#ifndef AUTOGRAPH_INIT_H
+#define AUTOGRAPH_INIT_H
+
+#ifdef __cplusplus
+extern "C" {
+#endif
+
+int autograph_init();
+
+#ifdef __cplusplus
+}  // extern "C"
+
+namespace autograph {
+
+bool init();
+
+}  // namespace autograph
+#endif
+
+#endif
