+++ cplusplus/src/core/init.c
@@ -1,4 +1,5 @@
-#include "internal.h"
+#include "init.h"
+
 #include "sodium.h"
 
 int autograph_init() { return sodium_init() >= 0 ? 0 : -1; }
