+++ cplusplus/src/core/sizes.c
@@ -1,4 +1,4 @@
-#include "internal.h"
+#include "sizes.h"
 
 unsigned int autograph_handshake_size() { return 80; }
