+++ cplusplus/src/init.cpp
@@ -1,0 +1,7 @@
+#include "init.h"
+
+namespace autograph {
+
+bool init() { return autograph_init() == 0; }
+
+}  // namespace autograph
